(PC-16753)[API] refactor: rename file to import_dms_applications
